fixed functionality to make and show reading lists
now database data can be seen in realtime when the data is changed, added button to make reading lists and show them in list format.

All of it needs styling I just created the basic functional component.

Pressing a list opens an alert for now, need change to show a page